+++ src/pushi/base/__main__.py
@@ -19,24 +19,18 @@
 # You should have received a copy of the Apache License along with
 # Hive Pushi System. If not, see <http://www.apache.org/licenses/>.
 
+__author__ = "João Magalhães <joamag@hive.pt>"
+""" The author(s) of the module """
+
 __copyright__ = "Copyright (c) 2008-2024 Hive Solutions Lda."
 """ The copyright for the module """
 
 __license__ = "Apache License, Version 2.0"
 """ The license for the module """
 
-from . import apn
-from . import app
-from . import base
-from . import event
-from . import subscription
-from . import web
-from . import web_push
+import pushi
 
-from .apn import APNAPI
-from .app import AppAPI
-from .base import BASE_URL, BASE_WS_URL, API
-from .event import EventAPI
-from .subscription import SubscriptionAPI
-from .web import WebAPI
-from .web_push import WebPushAPI
+state = pushi.State()
+app = pushi.PushiApp(state)
+server = pushi.PushiServer(state)
+state.load(app, server)